Version Update
This is the schematics of the update version (2.0); This update primarily focuses on integrating the necessary components into a single printed circuit board (PCB), making it extra compact and easy for maintenance. The fundamental electrical principles remain largely unchanged. For detailed information, please refer to the paper:
 Liu, Zhi and Suzuki, Shieru and Fushimi, Tatsuki and Ochiai, Yoichi, Voltage Controlled Howland Current Source for Galvanic Vestibular Stimulation. Available at SSRN: https://ssrn.com/abstract=5101503 or http://dx.doi.org/10.2139/ssrn.5101503 
Due to the potential risks associated with the use of the GVS device, please ensure that the device is functioning properly before use (see the README for details).